fix: refactored the smell of Bidirectional association
